+++ cli/cli/commands/test/testing_machinery/test_suite_runner/test_suite_runner_test.go
@@ -20,7 +20,6 @@ func TestBlockedExecutionWhenNoPerms(t *testing.T) {
 	result, err := RunTests(
 		perms,
 		nil,
-		nil,
 		logrus.InfoLevel,
 		"",
 		suiteMetadata,
@@ -40,7 +39,6 @@ func TestBlockedExecutionWhenRestrictedPerms(t *testing.T) {
 	})
 	result, err := RunTests(
 		perms,
-		nil,
 		nil,
 		logrus.InfoLevel,
 		"",